+++ db/migrate/20111213203218_create_invitations.rb
@@ -1,8 +1,8 @@
 class CreateInvitations < ActiveRecord::Migration
   def change
     create_table :invitations do |t|
-      t.integer :proposal_id
-      t.integer :facebook_user_id
+      t.bigint :proposal_id
+      t.bigint :facebook_user_id
       t.string :rsvp, :default => "Maybe"
       t.string :facebook_request_id
